M.k

See merge request internal/mypets!29

+++ app/Http/Controllers/API/ProfileController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API;
 
 use App\Components\Classes\StoreFile\File;
+use App\Http\Requests\API\ProfileUpdateRequest;
 use App\Http\Resources\UserResource;
 use App\Pet;
 use App\Picture;
@@ -33,10 +34,10 @@ class ProfileController extends Controller
     /**
      * Handle the incoming request.
      *
-     * @param Request $request
+     * @param ProfileUpdateRequest $request
      * @return Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(ProfileUpdateRequest $request)
     {
         $this->request = $request;
         $this->user = $request->user();
@@ -948,10 +949,10 @@ class ProfileController extends Controller
      */
     private function petPicturesHandler(Pet $pet, $pictures)
     {
-        if(isset($pictures['create']))
+        if(isset($pictures['create']) && count($pictures['create']))
             $this->savePetPictures($pet, $pictures['create']);
 
-        if(isset($pictures['delete']))
+        if(isset($pictures['delete']) && count($pictures['delete']))
             $this->destroyPetPictures($pet, $pictures['delete']);
     }
 
@@ -965,6 +966,9 @@ class ProfileController extends Controller
         $pictureModels = [];
 
         foreach ($pictures as $picture) {
+            if(!isset($picture))
+                continue;
+
             $file = $this->makeFile($picture, 'pictures');
             if($file)
                 $pictureModels[] = new Picture(['picture' => $file]);
